add adding the task functionality
