+++ src/classConstructor.js
@@ -1,0 +1,9 @@
+class Task {
+  constructor(value, index) {
+    this.description = value,
+    this.completed = false,
+    this.index = index
+  }
+}
+
+export default Task
