+++ src/createTask.js
@@ -1,18 +1,21 @@
 const taskInput = document.querySelector('.task-input')
-const taskSubmit = document.querySelector('.submit-task')
-const form =  document.querySelector('form')
-import { tasks } from "./index"
+const form = document.querySelector('form')
+let tasks = JSON.parse(localStorage.getItem('storageTasks'));
+tasks = (tasks === null) ? [] :  tasks
+import addTask from './addTask';
+import Task from './classConstructor';
 
-class Task {
-  constructor(value, index) {
-    this.description = value,
-    this.completed = false,
-    this.index = index
-  }
+function submit() {
+  form.addEventListener('submit', (e) => {
+    e.preventDefault()
+    if (taskInput.value !== '') {
+      const task = new Task(taskInput.value, tasks.length)
+      tasks.push(task)
+      localStorage.setItem('storageTasks', JSON.stringify(tasks));
+      addTask(task)
+      taskInput.value = ''
+    }
+  })
 }
 
-form.addEventListener('submit', () => {
-  if (taskInput.value !== '') {
-    const task = new Task(taskInput.value, tasks.length-1) 
-  }
-})
+export default submit
