+++ src/index.html
@@ -15,7 +15,7 @@
     </p>
     <form action="">
       <input class="task-input" type="text" placeholder="Add to your list...">
-      <button class="submit-task" type="button"><i class="fa-solid fa-arrow-left"></i></button>
+      <button class="submit-task"><i class="fa-solid fa-arrow-left"></i></button>
     </form>
     <ul id="todo-list">
     </ul>
